Fixed the `do-not-populate` utility generator
Fixes JITX-5565

The ocdb library had a function `do-not-populate`. Unfortanutely, there
was also a jstanza macro definition by this same name.  This caused the
symbol `do-not-populate` to be converted to a `set-do-not-populate`. The
result  was that the ocdb function was never called.

I've renamed `ocdb/utils/generator-utils/do-not-populate` to be `dnp` and
it will now handle any Instance or Tuple<Instance>.